move: Discount, EventBadge, Gift 클래스 domain.event 패키지로 이동
