+++ src/main/kotlin/christmas/domain/Receipt.kt
@@ -1,5 +1,6 @@
 package christmas.domain
 
+import christmas.domain.event.*
 import christmas.domain.menu.Menu
 
 class Receipt(private val visitDate: VisitDate, private val menu: Menu) { // todo 이름변경
