+++ src/test/kotlin/christmas/domain/event/DiscountTest.kt
@@ -1,5 +1,6 @@
-package christmas.domain
+package christmas.domain.event
 
+import christmas.domain.VisitDate
 import christmas.domain.menu.Dessert
 import christmas.domain.menu.MainDish
 import christmas.domain.menu.Menu
